+++ server/src/index.ts
@@ -2,6 +2,7 @@ import { ApolloServer } from 'apollo-server';
 import { initData } from './data';
 import { getResolvers, gqlTypeDefs } from './graphql';
 import 'dotenv/config';
+import logger from './logger';
 
 async function main() {
   const dataApi = await initData();
@@ -9,7 +10,8 @@ async function main() {
 
   const server = new ApolloServer({ typeDefs: gqlTypeDefs, resolvers });
   server.listen().then(({ url }) => {
-    console.log(`Server ready at ${url}`);
+    logger.info(`Server ready at ${url}`);
+    // console.log(`Server ready at ${url}`);
   });
 }